Add parameter handling options when generating SQL
While developing a native Oracle client (ODP.NET) adapter for FastDAO I
needed
to have more control over the way FastDAO handles parameter placeholder
naming.

To handle the spatial queries correctly, I needed to be able to wrap the
standard ? with a text conversion e.g. ST_GeomFromText(?,2272). The
new valueStrings dictionary allows for setting a custom parameter
placeholder.

The native Oracle client also requires the use of :param style
placeholders
rather than ?. The XML documentation said that this was supported, but I
discovered that there was no place where the ? were being converted in
the sql statement itself. To fix this I added a
FindParameterPlaceholders
method to SqlUtilities to locate the ?s, which can be used by a FastDAO
connection descriptor in the SetParametersOnCommand method