Add backwards compatible constructor for SortOrder
SortOrder was recently extended to accept a defulat-null parameter
called `parameters` which were values to be substituted into the
query string in place of ??? tokens.  Even though the parameter
was null by default, the signature was changed and dependent
libraries could no longer find a matching signature, requiring
everything to be updated and recompiled and redeployed to nuget.
Instead, I'm adding a traditional overloaded constructor that
should be backwards compatible with previous versions of the
library.